functions/requirements.txt
firebase-functions
flask
requests
google-api-python-client
google-auth
google-auth-httplib2
dotenv